Added ovs bridge configuration for 1 port only. Modified mlnx_bf_configure to create ovs bridge on DPU with the next interfaces: pf0hpf - pf represntor p0 - uplink representor en1f0pf0sf0 - SF
This change only support DPU with one port.
When there will be nulti port DPU support the script must be modified.

Starting openvswitch-switch service was inserted to this script
as a workaroud because this service was not running when the script
is executed.

Signed off-by: Emanuel Alkobi <ealkobi@nvidia.com>